uuid: explicitly cast floor() parameter to a float in GlobalIdGenerator
In PHP 8, this function is stricter about its input,
and warns when it is given a string, such as the case here.

Bug: T248925

+++ includes/libs/uuid/GlobalIdGenerator.php
@@ -398,7 +398,7 @@ class GlobalIdGenerator {
 			}
 			// Fetch the counter value and increment it...
 			rewind( $handle );
-			$counter = floor( trim( fgets( $handle ) ) ) + $count; // fetch as float
+			$counter = floor( (float)trim( fgets( $handle ) ) ) + $count; // fetch as float
 			// Write back the new counter value
 			ftruncate( $handle, 0 );
 			rewind( $handle );
